TradePostService 기능 생성 및 테스트 코드 (#147)

+++ api/src/main/java/com/hcs/mapper/TradePostMapper.java
@@ -14,7 +14,7 @@ public interface TradePostMapper {
 
     int countByTitle(String title);
 
-    long insertTradePost(TradePost tradePost);
+    int insertTradePost(TradePost tradePost);
 
     int updateTradePost(TradePost tradePost);
